+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Bugfix
 * fix line wrapping when calling multiple echo `newline:false` [#952](https://github.com/jcubic/jquery.terminal/issues/952)
 * fix cursor position when echo `newline:false` fill the line
+* fix running typing animation on empty terminal [#950](https://github.com/jcubic/jquery.terminal/issues/950)
 
 ## 2.42.0
 ### Features